+++ examples/7.5.2-plugin-buffereddelete.php
@@ -1,4 +1,5 @@
 <?php
+
 require_once(__DIR__.'/init.php');
 
 use Solarium\Plugin\BufferedDelete\Event\Events;