+++ examples/7.6-plugin-prefetchiterator.php
@@ -1,4 +1,5 @@
 <?php
+
 require_once(__DIR__.'/init.php');
 
 htmlHeader();
@@ -11,7 +12,7 @@ $query = $client->createSelect();
 $query->setFields(array('id'));
 
 // cursor functionality can be used for efficient deep paging (since Solr 4.7)
-$query->setCursormark('*');
+$query->setCursorMark('*');
 // cursor functionality requires a sort containing a uniqueKey field as tie breaker on top of your desired sorts for the query
 $query->addSort('id', $query::SORT_ASC);