+++ tests/Integration/AbstractTechproductsTest.php
@@ -2826,10 +2826,10 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertSame(32, $i);
     }
 
-    public function testPrefetchIteratorWithCursormark()
+    public function testPrefetchIteratorWithCursorMark()
     {
         $select = self::$client->createSelect();
-        $select->setCursormark('*');
+        $select->setCursorMark('*');
         $select->addSort('id', SelectQuery::SORT_ASC);
         /** @var PrefetchIterator $prefetch */
         $prefetch = self::$client->getPlugin('prefetchiterator');
@@ -2847,7 +2847,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertSame(32, $i);
     }
 
-    public function testPrefetchIteratorWithoutAndWithCursormark()
+    public function testPrefetchIteratorWithoutAndWithCursorMark()
     {
         $select = self::$client->createSelect();
         $select->addSort('id', SelectQuery::SORT_ASC);
@@ -2862,7 +2862,7 @@ abstract class AbstractTechproductsTest extends TestCase
         }
 
         $select = self::$client->createSelect();
-        $select->setCursormark('*');
+        $select->setCursorMark('*');
         $select->addSort('id', SelectQuery::SORT_ASC);
         $prefetch->setQuery($select);
 
@@ -3355,6 +3355,23 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertCount(0, $result);
     }
 
+    /**
+     * Get the options to use for ManagedResources Exists commands.
+     *
+     * @return array
+     */
+    public function getManagedResourcesExistsCommandOptions(): array
+    {
+        // Solr 7 can use HEAD requests because it's unaffected by SOLR-15116 and SOLR-16274
+        if (7 === self::$solrVersion) {
+            return [
+                'useHeadRequest' => true,
+            ];
+        }
+
+        return [];
+    }
+
     public function testManagedStopwords()
     {
         /** @var StopwordsQuery $query */
@@ -3363,7 +3380,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $term = 'managed_stopword_test';
 
         // Check that stopword list exists
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertTrue($result->getWasSuccessful());
@@ -3376,7 +3393,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that added stopword exists
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $exists->setTerm($term);
         $query->setCommand($exists);
         $result = self::$client->execute($query);
@@ -3404,7 +3421,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that added stopword is gone
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $exists->setTerm($term);
         $query->setCommand($exists);
         $result = self::$client->execute($query);
@@ -3434,7 +3451,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $query->setName($name.uniqid());
 
         // Check that stopword list doesn't exist
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertFalse($result->getWasSuccessful());
@@ -3455,7 +3472,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that stopword list was created
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertTrue($result->getWasSuccessful());
@@ -3474,7 +3491,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that added stopword exists in its original lowercase form
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $exists->setTerm($term);
         $query->setCommand($exists);
         $result = self::$client->execute($query);
@@ -3493,7 +3510,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that stopword list is gone
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertFalse($result->getWasSuccessful());
@@ -3513,7 +3530,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $term = 'managed_synonyms_test';
 
         // Check that synonym map exists
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertTrue($result->getWasSuccessful());
@@ -3529,7 +3546,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that added synonym mapping exsists
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $exists->setTerm($term);
         $query->setCommand($exists);
         $result = self::$client->execute($query);
@@ -3571,7 +3588,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that added synonym mapping is gone
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $exists->setTerm($term);
         $query->setCommand($exists);
         $result = self::$client->execute($query);
@@ -3601,7 +3618,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $query->setName($name.uniqid());
 
         // Check that synonym map doesn't exist
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertFalse($result->getWasSuccessful());
@@ -3623,7 +3640,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that synonym map was created
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertTrue($result->getWasSuccessful());
@@ -3646,7 +3663,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that synonym exists in its original lowercase form
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $exists->setTerm($term);
         $query->setCommand($exists);
         $result = self::$client->execute($query);
@@ -3665,7 +3682,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertTrue($result->getWasSuccessful());
 
         // Check that synonym map is gone
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertFalse($result->getWasSuccessful());
@@ -3789,7 +3806,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertFalse($result->getWasSuccessful(), 'Check if SOLR-6853 is fixed.');
 
         // The resource still exists
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $query->setCommand($exists);
         $result = self::$client->execute($query);
         $this->assertTrue($result->getWasSuccessful());
@@ -3841,7 +3858,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertFalse($result->getWasSuccessful(), 'Check if SOLR-6853 is fixed.');
 
         // The term still exists
-        $exists = $query->createCommand($query::COMMAND_EXISTS);
+        $exists = $query->createCommand($query::COMMAND_EXISTS, $this->getManagedResourcesExistsCommandOptions());
         $exists->setTerm($term);
         $query->setCommand($exists);
         $result = self::$client->execute($query);