+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+
+## [6.2.4] 
+### Added
 - Symfony 6 support
 - Solr 9 support
 - Unified Highlighter support + improved support for other highlighters
@@ -15,9 +25,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adding nested child documents through `Document::setField()` and `Document::addField()`
 - Solarium\Component\Highlighting\Highlighting::setMethod() didn't set the correct request parameter
 
+### Changed
+- Solarium\QueryType\Select\Query\Query::setCursormark() and getCursormark() are now setCursorMark() and getCursorMark() with uppercase M
+- Managed resources execute GET requests for the Exists command by default to avoid SOLR-15116 and SOLR-16274. Set the 'useHeadRequest' option to `true` to execute HEAD requests instead.
+
 ### Removed
 - Solarium\QueryType\Stream\Expression, use Solarium\QueryType\Stream\ExpressionBuilder instead
 - Solarium\QueryType\Update\Query\Document::setFilterControlCharacters(), extend Update\Query\Query to use a custom request builder & helper if you don't want control characters filtered
+
 
 ## [6.2.3]
 ### Added
@@ -253,7 +268,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AdapterInterface` does not extend `ConfigurableInterface` anymore
 - `Http` Adapter does not implement `ConfigurableInterface` anymore
 - `Psr18Adapter` does not implement `ConfigurableInterface` anymore
-- Solarium Client now accepts any PSR-15 compatible event dispatcher (previously it had to be symfony's event dispatcher)
+- Solarium Client now accepts any PSR-14 compatible event dispatcher (previously it had to be the Symfony EventDispatcher)
 
 ### Removed
 - Zend2HttpAdapter